move Metrics definition from mod.rs to metered.rs; use metered storage in runtime; track inner Blob state to gracefully handle drop/close
