+++ runtime/src/storage/tokio.rs
@@ -1,6 +1,5 @@
 use crate::Error;
 use commonware_utils::{from_hex, hex};
-use prometheus_client::registry::Registry;
 use std::sync::Arc;
 use std::{io::SeekFrom, path::PathBuf};
 use tokio::{
@@ -8,8 +7,6 @@ use tokio::{
     io::{AsyncReadExt, AsyncSeekExt, AsyncWriteExt},
     sync::Mutex as AsyncMutex,
 };
-
-use super::Metrics;
 
 #[derive(Clone)]
 pub struct Config {
@@ -24,55 +21,32 @@ impl Config {
 
 #[derive(Clone)]
 pub struct Storage {
-    metrics: Arc<Metrics>,
     fs: Arc<AsyncMutex<()>>,
     cfg: Config,
 }
 
 impl Storage {
-    pub fn new(metrics: &mut Registry, cfg: Config) -> Self {
+    pub fn new(cfg: Config) -> Self {
         Self {
-            metrics: Arc::new(Metrics::new(metrics)),
             fs: AsyncMutex::new(()).into(),
             cfg,
         }
     }
 }
 
+#[derive(Clone)]
 pub struct Blob {
-    metrics: Arc<Metrics>,
     partition: String,
     name: Vec<u8>,
     file: Arc<AsyncMutex<(fs::File, u64)>>,
 }
 
 impl Blob {
-    fn new(
-        metrics: Arc<Metrics>,
-        partition: String,
-        name: &[u8],
-        file: fs::File,
-        len: u64,
-    ) -> Self {
-        metrics.open_blobs.inc();
+    fn new(partition: String, name: &[u8], file: fs::File, len: u64) -> Self {
         Self {
-            metrics,
             partition,
             name: name.into(),
             file: Arc::new(AsyncMutex::new((file, len))),
-        }
-    }
-}
-
-impl Clone for Blob {
-    fn clone(&self) -> Self {
-        // We implement `Clone` manually to ensure the `open_blobs` gauge is updated.
-        self.metrics.open_blobs.inc();
-        Self {
-            metrics: self.metrics.clone(),
-            partition: self.partition.clone(),
-            name: self.name.clone(),
-            file: self.file.clone(),
         }
     }
 }
@@ -110,13 +84,7 @@ impl crate::Storage for Storage {
         let len = file.metadata().await.map_err(|_| Error::ReadFailed)?.len();
 
         // Construct the blob
-        Ok(Blob::new(
-            self.metrics.clone(),
-            partition.into(),
-            name,
-            file,
-            len,
-        ))
+        Ok(Blob::new(partition.into(), name, file, len))
     }
 
     async fn remove(&self, partition: &str, name: Option<&[u8]>) -> Result<(), Error> {
@@ -184,8 +152,6 @@ impl crate::Blob for Blob {
             .read_exact(buf)
             .await
             .map_err(|_| Error::ReadFailed)?;
-        self.metrics.storage_reads.inc();
-        self.metrics.storage_read_bytes.inc_by(buf.len() as u64);
         Ok(())
     }
 
@@ -206,8 +172,6 @@ impl crate::Blob for Blob {
         if max_len > file.1 {
             file.1 = max_len;
         }
-        self.metrics.storage_writes.inc();
-        self.metrics.storage_write_bytes.inc_by(buf.len() as u64);
         Ok(())
     }
 
@@ -242,11 +206,5 @@ impl crate::Blob for Blob {
             .shutdown()
             .await
             .map_err(|_| Error::BlobCloseFailed(self.partition.clone(), hex(&self.name)))
-    }
-}
-
-impl Drop for Blob {
-    fn drop(&mut self) {
-        self.metrics.open_blobs.dec();
     }
 }
